# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for telemetry service",
        "title": "Telemetry API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/telemetry/v1/geofences": {
            "get": {
                "description": "Retrieve a paginated list of geofences with optional filtering by space and active status. Organization is resolved from X-Organization header or hostname (e.g., {org}.localhost)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geofences"
                ],
                "summary": "Get geofences",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Filter by active status",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new geofence with polygon geometry for a space. Organization is resolved from X-Organization header or hostname (e.g., {org}.localhost)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geofences"
                ],
                "summary": "Create geofence",
                "parameters": [
                    {
                        "description": "Geofence configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateGeofenceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreateGeofenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/telemetry/v1/geofences/{geofence_id}": {
            "get": {
                "description": "Retrieve a single geofence by its UUID. Organization is resolved from X-Organization header or hostname (e.g., {org}.localhost)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geofences"
                ],
                "summary": "Get geofence by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Geofence UUID",
                        "name": "geofence_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GeofenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing geofence by ID. Organization is resolved from X-Organization header or hostname (e.g., {org}.localhost)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geofences"
                ],
                "summary": "Update geofence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Geofence UUID",
                        "name": "geofence_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Geofence updates",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateGeofenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateGeofenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a geofence by ID. Organization is resolved from X-Organization header or hostname (e.g., {org}.localhost)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geofences"
                ],
                "summary": "Delete geofence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Geofence UUID",
                        "name": "geofence_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteGeofenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.PaginatedResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "results": {}
            }
        },
        "models.CreateGeofenceRequest": {
            "type": "object",
            "required": [
                "color",
                "features",
                "name",
                "type_zone"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "definition": {
                    "type": "object"
                },
                "features": {
                    "description": "GeoJSON features",
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "space_id": {
                    "type": "string"
                },
                "type_zone": {
                    "description": "\"type_zone\" from frontend",
                    "type": "string",
                    "enum": [
                        "safe",
                        "danger",
                        "normal"
                    ]
                }
            }
        },
        "models.CreateGeofenceResponse": {
            "type": "object",
            "properties": {
                "geofence": {
                    "$ref": "#/definitions/models.GeofenceResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.DeleteGeofenceResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.GeofenceResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "event_rule": {
                    "$ref": "#/definitions/models.EventRuleInfo"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "space_id": {
                    "type": "string"
                },
                "type_zone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateGeofenceResponse": {
            "type": "object",
            "properties": {
                "geofence": {
                    "$ref": "#/definitions/models.GeofenceResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "API Key": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "Access Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Organization slug name": {
            "type": "apiKey",
            "name": "X-Organization",
            "in": "header"
        },
        "Space slug name": {
            "type": "apiKey",
            "name": "X-Space",
            "in": "header"
        }
    }
}